requirements.txt
carputils==0.0.0
contourpy==1.3.2
cycler==0.12.1
EditorConfig==0.17.1
fonttools==4.61.1
h5py==3.15.1
intel-cmplr-lib-ur==2025.3.2
intel-openmp==2025.3.2
joblib==1.5.3
jsbeautifier==1.15.4
json5==0.12.0
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.7
mkl==2025.1.0
msgpack==1.1.0
ndindex==1.10.0
netgen-mesher==6.2.2601
netgen_occt==7.8.1
ngsolve==6.2.2601
numpy==1.26.4
onemkl-license==2025.3.1
packaging==26.0
pillow==12.1.1
platformdirs==4.9.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-learn==1.7.2
scipy==1.11
six==1.14.0
tbb==2022.3.1
tcmlib==1.4.1
threadpoolctl==3.6.0
tomli==2.4.0
tqdm==4.67.3
umf==1.0.3
yapf==0.43.0
